% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{contains}
\alias{contains}
\title{Check if list contains a name}
\usage{
contains(s, L)
}
\arguments{
\item{s}{a string name}

\item{L}{a list object}
}
\value{
boolean
}
\description{
Check if list contains a name
}
