% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimfilter-package.R
\docType{package}
\name{kimfilter-package}
\alias{kimfilter}
\alias{kimfilter-package}
\title{kimfilter: Kim Filter}
\description{
'Rcpp' implementation of the multivariate Kim filter, which combines the Kalman and Hamilton filters for state probability inference. The filter is designed for state space models and can handle missing values and exogenous data in the observation and state equations. Kim, Chang-Jin and Charles R. Nelson (1999) "State-Space Models with Regime Switching: Classical and Gibbs-Sampling Approaches with Applications" \doi{10.7551/mitpress/6444.001.0001}\url{http://econ.korea.ac.kr/~cjkim/}.
}
\author{
\strong{Maintainer}: Alex Hubbard \email{hubbard.alex@gmail.com}

}
\keyword{internal}
