% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{kimfilter}
\alias{kimfilter}
\title{Kim Filter}
\description{
\emph{kimfilter} Rcpp implementation of the multivariate Kim filter, which 
combines the Kalman and Hamilton filters for state probability inference. 
The filter is designed for state space models and can handle missing values 
and exogenous data in the observation and state equations.
\code{browseVignettes("kimfilter")} to view it in your browser.
}
\author{
Alex Hubbard
}
