% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rginv}
\alias{Rginv}
\title{R's implementation of the Moore-Penrose pseudo matrix inverse}
\usage{
Rginv(m)
}
\arguments{
\item{m}{matrix}
}
\value{
matrix inverse of m
}
\description{
R's implementation of the Moore-Penrose pseudo matrix inverse
}
